requirements.txt
streamlit
requests
pydantic
beautifulsoup4
google-genai
python-dotenv
